requirements.txt
matplotlib==3.5.1
numpy==1.22.0
pandas==1.4.2
Pillow==9.0.1
Pillow==10.0.0
rpy2==3.5.13
scikit_learn==1.0.2
streamlit==1.14.0
tensorflow==2.13.0
